requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pathlib2>=2.3.7; python_version < "3.4"
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-aws>=0.1.0
